third_party/blink/web_tests/css-parser/deprecated-css-appearance-values-warning/appearance-with-nonstandard-keywords.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
/* See https://drafts.csswg.org/css-ui-4/#appearance-switching */
/* TODO(crbug.com/40261389): this test should be deleted once slider-vertical
   is removed. */
.t {
  appearance: auto;
  appearance: checkbox;
  appearance: radio;
  appearance: push-button;
  appearance: square-button;
  appearance: button;
  appearance: inner-spin-button;
  appearance: listbox;
  appearance: media-slider;
  appearance: media-sliderthumb;
  appearance: media-volume-slider;
  appearance: media-volume-sliderthumb;
  appearance: menulist;
  appearance: menulist-button;
  appearance: meter;
  appearance: progress-bar;
  appearance: slider-horizontal;
  appearance: slider-vertical;
  appearance: sliderthumb-horizontal;
  appearance: sliderthumb-vertical;
  appearance: searchfield;
  appearance: searchfield-cancel-button;
  appearance: textfield;
  appearance: textarea;
}
</style>
<script>
test(() => {
  assert_true(false, 'This should FAIL so we generate a -expected.txt');
}, 'There should be console warnings about non-standard appearance keywords');
</script>
